Zwischenstand

+++ 20Erfassen/Erfassen.py
@@ -18,6 +18,7 @@ from dbparam import *
 from suchen import suchen
 from bereinigen import bereinigen
 from statistik import statistik
+from mountlogs import mountLogs, unmountLogs
 
 TITEL = "Erfassen"
 VERSION = "V0"
@@ -28,18 +29,31 @@ mit neuen Daten erweitern
 """
 
 
-def main(pfad, neu=False, Dbg=False):
+def main(pfad, keep=True, neu=False, Dbg=False):
     try:
         mydb = mysql.connector.connect(
             host=DBHOST, db=DBNAME, user=DBUSER, port=DBPORT, password=DBPWD
         )
-        if neu:
-            with mydb.cursor() as cursor:
-                cursor.execute(f"TRUNCATE TABLE {DBTMW};")
         if pfad is not None:
+            pfad = mountLogs(pfad)
+            if pfad is None:
+                # mount-Fehler
+                logging.error("Logs konnten nicht eingebunden werden")
+                return "Fehler"
+            # else:
+            # es wird nur gelöscht, wenn auch neue Daten da sind
+            if neu:
+                with mydb.cursor() as cursor:
+                    cursor.execute(f"TRUNCATE TABLE {DBTMW};")
+            # suchen und DB füllen
             suchen(pfad, mydb, Dbg)
+        # egal, ob Pfad oder nicht:
+        # doppelte Einträge usw.
         bereinigen(mydb, Dbg)
+        # Überblick:
         statistik(mydb, Dbg)
+        # keine weitere Verarbeitung:
+        # jetzt stehen die Daten in der DB zur Verfügung
     except mysql.connector.errors.ProgrammingError as e:
         logging.error(e)
         match e.errno:
@@ -53,6 +67,8 @@ def main(pfad, neu=False, Dbg=False):
         logging.exception(e)
     finally:
         mydb.close()
+        if not keep:
+            unmountLogs(pfad)
 
     return 0
 
@@ -78,6 +94,13 @@ if __name__ == "__main__":
         help="DB leeren und neu erfassen (sonst nur ergänzen)",
     )
     parser.add_argument(
+        "-k",
+        "--keep",
+        dest="pKeep",
+        action="store_true",
+        help="Logfiles behalten (kein Un-Mount nach Verarbeitung)",
+    )
+    parser.add_argument(
         "pfad",
         nargs="?",
         default=None,
@@ -88,10 +111,11 @@ if __name__ == "__main__":
     pfad = arguments.pfad
     Dbg = arguments.pVerbose
     neu = arguments.pNeu
+    keep = arguments.pKeep
     if Dbg:
         LOG_LEVEL = logging.DEBUG
     else:
         LOG_LEVEL = logging.INFO
     logging.basicConfig(format=LOG_FORMAT, level=LOG_LEVEL)
 
-    sys.exit(main(pfad, neu, Dbg))
+    sys.exit(main(pfad, keep, neu, Dbg))
